requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
scikit-learn==1.4.0
scipy==1.11.4
joblib==1.3.2
beautifulsoup4==4.12.3
lxml==5.1.0
numpy==1.26.3
pydantic==2.6.0
gdown==5.1.0
requests==2.31.0